mtk3: LP64 ホスト対応の最小パッチ（コア共通部）
μT-Kernel 3.0 コアを LP64（Linux x86-64 ユーザモード）で動かすための
上流互換パッチ。32bit MCU ターゲットでは動作は従来と同一。

- knlinc/kernel.h: <config.h> を先頭で include。元の並びでは
  <tk/typedef.h> 処理時に USE_STDINC_STDINT が未定義となり、
  tk/tkernel.h 経由の TU と型幅（W/UW）が食い違う
- tstdlib/string.c: knl_memset のワードループが unsigned long
  （LP64 で 8 バイト）で書きながら n を 4 ずつしか減らさず、
  要求の約 2 倍を塗り潰して呼び出し元スタックを破壊していた。
  整列・埋め値・減算を sizeof(unsigned long) 基準に統一
- memory.c / memory.h / mempool.c: ポインタ演算の (UW) キャストを
  unsigned long に変更（LP64 では上位 32bit が切り捨てられる）
- inittask.h: 初期タスクパラメータを #ifndef ガード化し
  ターゲット側から上書き可能に

+++ kernel/mtkernel3/kernel/tkernel/mempool.c
@@ -99,7 +99,7 @@ LOCAL void knl_appendFreeAreaBound( MPLCB *mplcb, QUEUE *aq )
 		/* FreeQue Size order */
 		QueInsert(top, fq);
 		(top + 1)->next = NULL;
-		(top + 1)->prev = (QUEUE*)size;
+		(top + 1)->prev = (QUEUE*)(unsigned long)size;	/* p-kernel: LP64 の警告回避（サイズ値の格納） */
 	}
 }
